refactor: rename education/schema.ts to types.ts
The file holds TypeScript interfaces for educational content (no DB
schema, no Zod, no SQL). The previous name false-tripped the
Schema Drift Gate keyword scanner, which treats any path containing
"schema" outside of mobile/web/docs/scripts as DB-impacting and demands
a matching ctf/schema.sql change.

Renames to types.ts and updates the three import sites + README +
inventory change log accordingly.
