+++ ctf/packages/education/index.ts
@@ -1,2 +1,2 @@
-export * from './schema';
+export * from './types';
 export * from './components/EducationModal';
